refactor: move http_method out of utils
