+++ pkg/http_method/http_method.go
@@ -1,0 +1,24 @@
+package http_method
+
+import (
+	"net/http"
+	"strings"
+)
+
+var methodMap = map[string]string{
+	"GET":    http.MethodGet,
+	"POST":   http.MethodPost,
+	"PUT":    http.MethodPut,
+	"PATCH":  http.MethodPatch,
+	"DELETE": http.MethodDelete,
+}
+
+func StringToHttpMethod(method string) string {
+	uppercaseMethod := strings.ToUpper(method)
+
+	if value, ok := methodMap[uppercaseMethod]; ok {
+		return value
+	}
+
+	return "UNKNOWN"
+}
